Factors out sports cpmmands. Unfinished.
